api/requirements.txt
django==4.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
